Fixed-#24641;#24640;#24639
git-svn-id: https://ncisvn.nci.nih.gov/svn/coppa/trunk/code@7794 553898ca-32d5-487a-af8b-6cf4ca579ab4

+++ pa/pa-ejb/src/java/gov/nih/nci/pa/service/util/TSRReportGeneratorServiceBean.java
@@ -786,7 +786,7 @@ private String getInterventionAltNames(InterventionDTO i) throws PAException {
                if (pq.getLow().getValue().intValue() == 0) {
                   html.append(appendData("Minimum Age", "N/A" , true , true));
                } else {
-                  html.append(appendData("Minimum Age", pq.getLow().getValue()  
+                  html.append(appendData("Minimum Age", pq.getLow().getValue().intValue()  
                   + " " + pq.getLow().getUnit(), true , true));
                }
              } 
@@ -794,7 +794,7 @@ private String getInterventionAltNames(InterventionDTO i) throws PAException {
                if (pq.getHigh().getValue().intValue() == MAX_AGE) {
                   html.append(appendData("Maximum Age", "N/A" , true , true));
                } else {
-                  html.append(appendData("Maximum Age", pq.getHigh().getValue()  
+                  html.append(appendData("Maximum Age", pq.getHigh().getValue().intValue()  
                   + " " + pq.getHigh().getUnit(), true , true));
                }
              }
